精灵图元渲染功能实现

+++ yokirender/src/main/java/com/xinlan/yokirender/core/YokiBit.java
@@ -5,8 +5,6 @@ package com.xinlan.yokirender.core;
  */
 public class YokiBit {
     public int textureId = -1;
-    public int srcWidth;//原始位图 宽度
-    public int srcHeight; //原始位图 高度
-
-
+    public float srcWidth;//原始位图 宽度
+    public float srcHeight; //原始位图 高度
 }//end class
